--- src/main/java/org/zalando/nakadi/exceptions/IllegalClientIdException.java
@@ -1,9 +1,0 @@
-package org.zalando.nakadi.exceptions;
-
-public class IllegalClientIdException extends RuntimeException {
-
-    public IllegalClientIdException(final String message) {
-        super(message);
-    }
-
-}
